backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "databases>=0.9.0",
    "fastapi[standard,stander]>=0.115.14",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
]
